+++ Sources/Views/Settings/TTS/TTSTransliterationTesterView.swift
@@ -40,7 +40,6 @@ struct TTSTransliterationTesterView: View {
     var body: some View {
         Form {
             engineSection
-            TTSIPAProbeSection()
             sentenceSection
             wordSection
             goldenSection